feat(utils.py): working proto drive dump
This is a working version of the .proto drive dump for all of the observations generated from a single replay. There may be an easier way to have the same functionality here. Having a proto message that is a collection of messages seems more natural to do in this case.