feat(sc2_combat_detector): try, except, graceful exception catching and logging
During processing one of the threads stopped simulating replays, I suspect (without verification) that this was caused by an exception on one of the replays. These very general try, except blocks should prevent that and allow for more graceful way of running the program. I am awayre of the potentially sub-optimal way of catching every Exception.